--- data/repositories.js
@@ -1,35 +1,0 @@
-const projects = {
-    org: 'OpenDevUFCG',
-    repositories: [
-        {
-            name: 'calluswhatyouwant/spotify-web-sdk',
-            difficulty: 'avancado',
-        },
-        {
-            name: 'calluswhatyouwant/musicritic',
-            difficulty: 'avancado',
-        },
-        {
-            name: 'paulojbleitao/pokedex',
-            difficulty: 'intermediario',
-        },
-        {
-            name: 'SubmiBot/SubmiBot',
-            difficulty: 'intermediario',
-        },
-        {
-            name: 'marianabianca/pomodoro',
-            difficulty: 'iniciante',
-        },
-        {
-            name: 'P-Sync/P-Sync',
-            difficulty: 'intermediario',
-        },
-        {
-            name: 'FelipeMarinho97/iskra-webkit-greeter',
-            difficulty: 'intermediario',
-        },
-    ],
-};
-
-export default projects;
