+++ data/repositories.json
@@ -1,0 +1,41 @@
+{
+    "org": "OpenDevUFCG",
+    "repositories": [
+        {
+            "name": "calluswhatyouwant/spotify-web-sdk",
+            "difficulty": "avancado"
+        },
+        {
+            "name": "calluswhatyouwant/musicritic",
+            "difficulty": "avancado"
+        },
+        {
+            "name": "paulojbleitao/pokedex",
+            "difficulty": "intermediario"
+        },
+        {
+            "name": "SubmiBot/SubmiBot",
+            "difficulty": "intermediario"
+        },
+        {
+            "name": "marianabianca/pomodoro",
+            "difficulty": "iniciante"
+        },
+        {
+            "name": "P-Sync/P-Sync",
+            "difficulty": "intermediario"
+        },
+        {
+            "name": "FelipeMarinho97/iskra-webkit-greeter",
+            "difficulty": "intermediario"
+        },
+        {
+            "name": "Rickecr/PyGraph",
+            "difficulty": "intermediario"
+        },
+        {
+            "name": "Rickecr/CamaraDosDeputados",
+            "difficulty": "iniciante"
+        }
+    ]
+}
